Moved BRDF loading to gl-wrangling-functions.
Also cleaned up commented out code that is no longer needed.
